fix:refactor dashboard ui files structure
