+++ src/components/features/dashboard/admin/admin-dashboard-view.tsx
@@ -8,19 +8,26 @@ import { TYPOGRAPHY_CLASSNAMES } from "@/components/shared/typography"
 import { DisposeAssetsRequestDialog, type SelectedAssetLite } from "@/components/features/disposals/dispose-assets-request-dialog"
 import { tiqriToast } from "@/components/shared/sonner"
 import { sendAssignmentReminderAction } from "@/actions/assignments"
-import { KpiMetricsRow } from "./shared/kpi-metrics-row"
-import { DepartmentAllocationChart } from "./shared/department-allocation-chart"
-import { InventoryStatusChart } from "./shared/inventory-status-chart"
-import { DataTablesContainer } from "./shared/data-tables-container"
-import { useOverdueColumns, useHighMaintenanceColumns } from "./shared/dashboard-table-columns"
-import type { ITDashboardBatchData } from "@/actions/dashboard/it-operator"
+import { KpiMetricsRow } from "../shared/kpi-metrics-row"
+import { DepartmentAllocationChart } from "../shared/department-allocation-chart"
+import { InventoryStatusChart } from "../shared/inventory-status-chart"
+import { RecentActivitiesList } from "../shared/recent-activities-list"
+import { DataTablesContainer } from "../shared/data-tables-container"
+import { useOverdueColumns, usePendingDisposalColumns, useHighMaintenanceColumns } from "../shared/dashboard-table-columns"
+import type { AdminDashboardBatchData } from "@/actions/dashboard/admin"
 import type { OverdueReturnRow, HighMaintenanceRow } from "@/actions/dashboard/shared"
 
-interface ITDashboardViewProps {
-  data: ITDashboardBatchData
+interface AdminDashboardViewProps {
+  data: AdminDashboardBatchData
+  currencyCode?: string
+  exchangeRate?: number
 }
 
-export function ITDashboardView({ data }: ITDashboardViewProps) {
+export function AdminDashboardView({
+  data,
+  currencyCode = "LKR",
+  exchangeRate = 1,
+}: AdminDashboardViewProps) {
   const [flaggedAsset, setFlaggedAsset] = useState<SelectedAssetLite | null>(null)
   const [isFlagDialogOpen, setIsFlagDialogOpen] = useState(false)
   const [sendingReminderIds, setSendingReminderIds] = useState<number[]>([])
@@ -51,6 +58,7 @@ export function ITDashboardView({ data }: ITDashboardViewProps) {
   }
 
   const overdueColumns = useOverdueColumns("Send Reminder", handleSendReminder, sendingReminderIds)
+  const pendingColumns = usePendingDisposalColumns("GlobalAdmin")
   const lemonsColumns = useHighMaintenanceColumns(handleFlagClick)
 
   const tableProps = {
@@ -78,6 +86,19 @@ export function ITDashboardView({ data }: ITDashboardViewProps) {
             {data.overdueReturns.length}
           </span>
         </TabsTrigger>
+        <TabsTrigger
+          value="pending"
+          className="group flex items-center gap-1.5 text-xs data-[state=active]:bg-background data-[state=active]:shadow-sm"
+        >
+          Pending Disposals
+          <span className={cn(
+            "text-[9px] font-semibold rounded-full px-1.5 py-0.5 leading-none transition-colors",
+            "group-data-[state=active]:bg-primary group-data-[state=active]:text-primary-foreground",
+            "group-data-[state=inactive]:bg-background group-data-[state=inactive]:text-primary border border-primary/30"
+          )}>
+            {data.pendingDisposals.length}
+          </span>
+        </TabsTrigger>
       </TabsList>
 
       <TabsContent value="overdue">
@@ -88,6 +109,17 @@ export function ITDashboardView({ data }: ITDashboardViewProps) {
           emptyState={{
             title: "No overdue returns",
             description: "All assets have been returned on time.",
+          }}
+        />
+      </TabsContent>
+      <TabsContent value="pending">
+        <DataTable
+          {...tableProps}
+          columns={pendingColumns}
+          data={data.pendingDisposals}
+          emptyState={{
+            title: "No pending disposals",
+            description: "There are no disposal requests awaiting review.",
           }}
         />
       </TabsContent>
@@ -116,21 +148,25 @@ export function ITDashboardView({ data }: ITDashboardViewProps) {
   return (
     <div className="px-6 py-1 pb-5 flex flex-col gap-6">
       {/* KPIs */}
-      <KpiMetricsRow metrics={data.kpiMetrics} />
+      <KpiMetricsRow
+        metrics={data.kpiMetrics}
+        currencyCode={currencyCode}
+        exchangeRate={exchangeRate}
+      />
 
       {/* Charts Grid */}
-      <div className="grid gap-4 min-h-[280px] grid-cols-1 lg:grid-cols-2">
+      <div className="grid gap-4 min-h-[280px] grid-cols-1 lg:grid-cols-3">
         <DepartmentAllocationChart allocationData={data.departmentAllocation} />
         <InventoryStatusChart
           inventoryData={data.inventoryStatus.inventoryData}
           utilizationRate={data.inventoryStatus.utilizationRate}
         />
+        <RecentActivitiesList activities={data.recentActivities} />
       </div>
 
       {/* Tables Container */}
       <DataTablesContainer leftSection={leftTables} rightSection={rightTables} />
 
-      {/* Disposal dialogue */}
       <DisposeAssetsRequestDialog
         open={isFlagDialogOpen}
         onOpenChange={setIsFlagDialogOpen}
